refactor : 폴더 구조 변경 (등록 관련 컴포넌트 공통 컴포넌트로 분리)
