+++ components/common/register/FormDetail.tsx
