+++ components/common/register/Thumbnail.tsx
